feat(entity-browser): add sorting to the react-redux-grid
The react-redux-grid component provides two different ways to sort the grid by their columns.
One way would be a `local` mode which just sorts the data which are currently visible in the grid.
The other way a `remote` sorting mode can be configured. In this mode an URL has to be set
which can be called to get the sorted data.
Both ways are not really what we want. We actually just want to set the state with the new ordering.
The data will then be set correctly via other actions following the `SET_ORDER` action.

And this is also how it gets done now. It's not mentioned in the documentation of the react-redux-grid
but a look into the implementation of its grid actions showed, that it is possible to provide a function
instead of a URL in the remote mode. This function gets the ordering as argument and should return
the sorted data by a promise object.

See: https://github.com/bencripps/react-redux-grid/blob/master/src/actions/GridActions.js

The list view already gets a function `setOrderBy` which we can call with the current ordering. To
satisfy react-redux-grid a promise with an empty object will be returned which leads to a js-console
warning which is not really a satisfying solution.